Fix ActiveChat not updating on new message.
- ActiveChat's old `mapStateToProps` function was returning values that
looked identical to React, even when there was a new message to render.
This was causing messages to "disappear" after being sent, and new
messages not to appear at all until the client was refreshed.
- The new `mapStateToProps` has a new attribute: `nMessages`,
representing the number of messages in the current conversation. This
value will go up when a new message is recieved - causing `ActiveChat`
to re-render.