Refactor two `reducerFunctions` into immutable functions.
- `addMessageToStore` and `addNewConvoToStore` were mutating the state,
causing ActiveChat and possibly other React components to fail to
re-render when Redux data had been changed.
- They have no both been refactored into immutable functions, resulting
in the correct behavior of UI elements re-rendering as-needed.
- This commit shouldn't have any side effects, unless existing code
relies on the side effects of `addMessageToStore` or
`addNewConvoToStore` - or on the lack of re-rendering that resulted from
the way these functions were written.